Compute verification scores for deterministic forecasts.

Compute verification scores for deterministic forecasts.

Usage

det_verify(
  .fcst,
  parameter,
  thresholds = NULL,
  groupings = "lead_time",
  circle = NULL,
  hexbin = TRUE,
  num_bins = 30,
  show_progress = TRUE,
  ...
)

# S3 method for harp_det_point_df
det_verify(
  .fcst,
  parameter,
  thresholds = NULL,
  groupings = "lead_time",
  circle = NULL,
  hexbin = TRUE,
  num_bins = 30,
  show_progress = TRUE,
  fcst_model = NULL,
  ...
)

Arguments

.fcst

A harp_df or harp_list object with tables that have a column for observations, or a single forecast table.

parameter

The name of the column for the observations data. Can be the column name, quoted, or unquoted. If a variable it should be embraced - i.e. wrapped in {{}}.

thresholds

A numeric vector of thresholds for which to compute the threshold based scores. Set to NULL (the default) to only compute summary scores.

groupings

The groups for which to compute the scores. See group_by for more information of how grouping works.

circle

If set the parameter is assumed to be cyclic for bias calculations. Should be this distance around a circle in the units of the parameter, so would typically have a value of 360 for degrees or 2 * pi for radians.

hexbin

Logical. Whether to compute hexbins for forecast, observation pairs. Defaults to TRUE. See bin_fcst_obs for more details.

num_bins

The number of bins into which to partition observations for the hexbin computation.

show_progress

Logical - whether to show progress bars. Defaults to TRUE.

...

Reserved for methods.

fcst_model

The name of the forecast model to use in the fcst_model column of the output. If the function is dispatched on a harp_list object, the names of the harp_list are automatically used.