Compute the skill (RMSE) and spread of an ensemble forecast

The ensemble mean and spread are computed as columns in a harp_list object. Typically the scores are aggregated over lead time by other grouping variables cam be chosen. The mean bias is also computed.

Usage

ens_spread_and_skill(
  .fcst,
  parameter,
  groupings = "lead_time",
  circle = NULL,
  spread_drop_member = NULL,
  jitter_fcst = NULL,
  show_progress = TRUE,
  ...
)

# S3 method for harp_ens_point_df
ens_spread_and_skill(
  .fcst,
  parameter,
  groupings = "lead_time",
  circle = NULL,
  spread_drop_member = NULL,
  jitter_fcst = NULL,
  show_progress = TRUE,
  fcst_model = NULL,
  ...
)

Arguments

.fcst

A harp_df or harp_list object with tables that have a column for observations, or a single forecast table.

parameter

The name of the column for the observations data. Can be the column name, quoted, or unquoted. If a variable it should be embraced - i.e. wrapped in {{}}.

groupings

The groups for which to compute the scores. See group_by for more information of how grouping works.

circle

If set the parameter is assumed to be cyclic for bias calculations. Should be this distance around a circle in the units of the parameter, so would typically have a value of 360 for degrees or 2 * pi for radians.

spread_drop_member

Which members to drop for the calculation of the ensemble variance and standard deviation. For harp_fcst objects, this can be a numeric scalar - in which case it is recycled for all forecast models; a list or numeric vector of the same length as the harp_fcst object, or a named list with the names corresponding to names in the harp_fcst object.

jitter_fcst

A function to perturb the forecast values by. This is used to account for observation error in the rank histogram. For other statistics it is likely to make little difference since it is expected that the observations will have a mean error of zero.

show_progress

Logical - whether to show progress bars. Defaults to TRUE.

...

Reserved for methods.

fcst_model

The name of the forecast model to use in the fcst_model column of the output. If the function is dispatched on a harp_list object, the names of the harp_list are automatically used.

Value

An object of the same format as the inputs but with data grouped for the groupings column(s) and columns for rmse, spread

and mean_bias.