Set or modify and output the perfect scores for different verificaiton metrics

The function returns a named vector giving the perfect score for all verification scores computed by harpPoint that give a single value. These perfect values can be modified by specifying them in named arguments and perfect scores for new scores can be set in the same way.

Usage

perfect_score(...)

Arguments

...

Named arguments that give values for the perfect score

Value

A named vector

Examples

perfect_score()
#>                     bias                     rmse                      mae 
#>                    0e+00                    0e+00                    0e+00 
#>                     stde             threat_score                 hit_rate 
#>                    0e+00                    1e+00                    1e+00 
#>                miss_rate         false_alarm_rate        false_alarm_ratio 
#>                    0e+00                    0e+00                    0e+00 
#>       heidke_skill_score       pierce_skill_score       kuiper_skill_score 
#>                    1e+00                    1e+00                    1e+00 
#>          percent_correct           frequency_bias   equitable_threat_score 
#>                    1e+00                    1e+00                    1e+00 
#>               odds_ratio           log_odds_ratio   odds_ratio_skill_score 
#>                    1e+06                    1e+06                    1e+00 
#> extreme_dependency_score            symmetric_eds extreme_dependency_index 
#>                    1e+00                    1e+00                    1e+00 
#>            symmetric_edi                mean_bias                   spread 
#>                    1e+00                    0e+00                    1e+06 
#>       spread_skill_ratio                     crps           crps_potential 
#>                    1e+00                    0e+00                    0e+00 
#>         crps_reliability         fair_brier_score                fair_crps 
#>                    0e+00                    0e+00                    0e+00 
#>              brier_score        brier_skill_score  brier_score_reliability 
#>                    0e+00                    1e+00                    0e+00 
#>   brier_score_resolution                 roc_area 
#>                    1e+00                    1e+00 
perfect_score(bss = 1)
#>                     bias                     rmse                      mae 
#>                    0e+00                    0e+00                    0e+00 
#>                     stde             threat_score                 hit_rate 
#>                    0e+00                    1e+00                    1e+00 
#>                miss_rate         false_alarm_rate        false_alarm_ratio 
#>                    0e+00                    0e+00                    0e+00 
#>       heidke_skill_score       pierce_skill_score       kuiper_skill_score 
#>                    1e+00                    1e+00                    1e+00 
#>          percent_correct           frequency_bias   equitable_threat_score 
#>                    1e+00                    1e+00                    1e+00 
#>               odds_ratio           log_odds_ratio   odds_ratio_skill_score 
#>                    1e+06                    1e+06                    1e+00 
#> extreme_dependency_score            symmetric_eds extreme_dependency_index 
#>                    1e+00                    1e+00                    1e+00 
#>            symmetric_edi                mean_bias                   spread 
#>                    1e+00                    0e+00                    1e+06 
#>       spread_skill_ratio                     crps           crps_potential 
#>                    1e+00                    0e+00                    0e+00 
#>         crps_reliability         fair_brier_score                fair_crps 
#>                    0e+00                    0e+00                    0e+00 
#>              brier_score        brier_skill_score  brier_score_reliability 
#>                    0e+00                    1e+00                    0e+00 
#>   brier_score_resolution                 roc_area                      bss 
#>                    1e+00                    1e+00                    1e+00 
perfect_score(bias = -1)
#>                     bias                     rmse                      mae 
#>                   -1e+00                    0e+00                    0e+00 
#>                     stde             threat_score                 hit_rate 
#>                    0e+00                    1e+00                    1e+00 
#>                miss_rate         false_alarm_rate        false_alarm_ratio 
#>                    0e+00                    0e+00                    0e+00 
#>       heidke_skill_score       pierce_skill_score       kuiper_skill_score 
#>                    1e+00                    1e+00                    1e+00 
#>          percent_correct           frequency_bias   equitable_threat_score 
#>                    1e+00                    1e+00                    1e+00 
#>               odds_ratio           log_odds_ratio   odds_ratio_skill_score 
#>                    1e+06                    1e+06                    1e+00 
#> extreme_dependency_score            symmetric_eds extreme_dependency_index 
#>                    1e+00                    1e+00                    1e+00 
#>            symmetric_edi                mean_bias                   spread 
#>                    1e+00                    0e+00                    1e+06 
#>       spread_skill_ratio                     crps           crps_potential 
#>                    1e+00                    0e+00                    0e+00 
#>         crps_reliability         fair_brier_score                fair_crps 
#>                    0e+00                    0e+00                    0e+00 
#>              brier_score        brier_skill_score  brier_score_reliability 
#>                    0e+00                    1e+00                    0e+00 
#>   brier_score_resolution                 roc_area 
#>                    1e+00                    1e+00